+++ oneflow/core/kernel/maximum_kernel.cpp
@@ -29,7 +29,7 @@ void MaximumKernel<device_type, T>::BackwardDataContent(
   const Blob* mask_blob = BnInOp2Blob("mask");
   Blob* out_diff_blob = BnInOp2Blob(GenDiffBn("out"));
   const int count = mask_blob->shape().elem_cnt();
-  for (int i = 0; i < this->kernel_conf().input_diff_bns().size(); ++i) {
+  for (size_t i = 0; i < this->kernel_conf().input_diff_bns().size(); ++i) {
     const std::string idbn = this->kernel_conf().input_diff_bns()[i];
     Blob* in_diff_blob = BnInOp2Blob(idbn);
     Memset<device_type>(ctx.device_ctx, in_diff_blob->mut_dptr(), 0,
